LayoutTests/accessibility/custom-elements/flowto-expected.txt

This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById("custom-1").role is "AXRole: AXCheckBox"
PASS platformValueForW3CName(accessibilityController.accessibleElementById("custom-1")) is "label 1 label 2"
PASS accessibilityController.accessibleElementById("custom-1").ariaFlowToElementAtIndex(0).role is "AXRole: AXButton"
PASS accessibilityController.accessibleElementById("custom-1").ariaFlowToElementAtIndex(0).title is "AXTitle: FlowTo1"
PASS accessibilityController.accessibleElementById("custom-1").ariaFlowToElementAtIndex(1).role is "AXRole: AXButton"
PASS accessibilityController.accessibleElementById("custom-1").ariaFlowToElementAtIndex(1).title is "AXTitle: FlowTo2"
PASS accessibilityController.accessibleElementById("custom-1").ariaFlowToElementAtIndex(2) is null
PASS accessibilityController.accessibleElementById("custom-2").role is "AXRole: AXCheckBox"
PASS platformValueForW3CName(accessibilityController.accessibleElementById("custom-2")) is "label 2"
PASS accessibilityController.accessibleElementById("custom-2").ariaFlowToElementAtIndex(0).role is "AXRole: AXButton"
PASS accessibilityController.accessibleElementById("custom-2").ariaFlowToElementAtIndex(0).title is "AXTitle: FlowTo2"
PASS accessibilityController.accessibleElementById("custom-2").ariaFlowToElementAtIndex(1) is null
PASS successfullyParsed is true

TEST COMPLETE
FlowTo1
FlowTo2
label 1
label 2
